refactor: extract StorageKeys to separate file and cleanup utilities
- Move StorageKeys class to lib/core/constants/storage_keys.dart
- Use StorageKeys constants in AuthInterceptor instead of hardcoded string
- Remove redundant isNullOrEmpty/isNotNullOrEmpty string extensions
- Simplify Formatters.duration() to use DurationX.formattedCompact extension
- Improve NotFoundException to include actual request path
